+++ tests/error_message.js
@@ -1,26 +1,26 @@
 module.exports = ({ test, assert, affirm, alike }) => [
     test("Assert quotes string in error", () => {
         affirm(test('', () =>
-            assert(true, 'true')).error,
-            (e) => e.includes(`[true] === ['true']`))
+            assert(true, 'true')).error.message,
+            (m) => m.includes(`[true] === ['true']`))
         affirm(test('', () =>
-            assert(0, '0')).error,
-            (e) => e.includes(`[0] === ['0']`))
+            assert(0, '0')).error.message,
+            (m) => m.includes(`[0] === ['0']`))
     })
     , test("Alike quotes string in error", () => {
         affirm(test('', () =>
-            alike(true, 'true')).error,
-            (e) => e.includes(`[true] === ['true']`))
+            alike(true, 'true')).error.message,
+            (m) => m.includes(`[true] === ['true']`))
         affirm(test('', () =>
-            alike(0, '0')).error,
-            (e) => e.includes(`[0] === ['0']`))
+            alike(0, '0')).error.message,
+            (m) => m.includes(`[0] === ['0']`))
     })
     , test("Affirm quotes string in error", () => {
         affirm(test('', () =>
-            affirm(true, 'true', false)).error,
-            (e) => e.includes('--> true\n') && e.includes(`--> 'true'\n`))
+            affirm(true, 'true', false)).error.message,
+            (m) => m.includes('--> true\n') && m.includes(`--> 'true'\n`))
         affirm(test('', () =>
-            affirm(0, '0', false)).error,
-            (e) => e.includes('--> 0\n') && e.includes(`--> '0'\n`))
+            affirm(0, '0', false)).error.message,
+            (m) => m.includes('--> 0\n') && m.includes(`--> '0'\n`))
     })
 ]